requirements.txt
streamlit>=1.31.0
google-generativeai>=0.4.1
faiss-cpu>=1.7.4
numpy>=1.24.0